runtime: ensure runtime is initialized
Make the `appinit` have no dependencies and call
out to the `encore.dev/appruntime/app` package
with `go:linkname` to actually initialize the app.
