+++ runtime/beta/auth/pkgfn.go
@@ -2,6 +2,11 @@
 
 package auth
 
+import (
+	// Ensure the runtime is initialized.
+	_ "encore.dev/appruntime/app/appinit"
+)
+
 //publicapigen:drop
 var Singleton *Manager // injected on app init
